# services/game/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GeoGuessr School Edition – game service",
        "title": "GSSR Game API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/rooms": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Create a new room",
                "parameters": [
                    {
                        "description": "Room settings",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/room.createRoomReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/rooms/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Get room state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/room.RoomState"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/rooms/{id}/guess": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Submit a guess for the current round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Guess coordinates",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/room.guessReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/rooms/{id}/join": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Join a room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/rooms/{id}/leave": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Leave a room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/rooms/{id}/start": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Start the game (host only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "room.Guess": {
            "type": "object",
            "properties": {
                "floor_id": {
                    "type": "string"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "room.Player": {
            "type": "object",
            "properties": {
                "elo": {
                    "type": "integer"
                },
                "has_guessed": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "room.RoomState": {
            "type": "object",
            "properties": {
                "current_pano_id": {
                    "type": "string"
                },
                "current_round": {
                    "type": "integer"
                },
                "guesses": {
                    "description": "Guesses maps userID → guess for the current round.",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/room.Guess"
                    }
                },
                "host_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "map_id": {
                    "type": "string"
                },
                "match_id": {
                    "type": "string"
                },
                "max_players": {
                    "type": "integer"
                },
                "pano_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/room.Player"
                    }
                },
                "round_token": {
                    "description": "RoundToken is regenerated each round to prevent double-advance races.",
                    "type": "string"
                },
                "rounds": {
                    "type": "integer"
                },
                "status": {
                    "description": "waiting|active|finished",
                    "type": "string"
                },
                "time_limit_sec": {
                    "type": "integer"
                }
            }
        },
        "room.createRoomReq": {
            "type": "object",
            "properties": {
                "map_id": {
                    "type": "string"
                },
                "max_players": {
                    "type": "integer"
                },
                "rounds": {
                    "type": "integer"
                },
                "time_limit_sec": {
                    "type": "integer"
                }
            }
        },
        "room.guessReq": {
            "type": "object",
            "properties": {
                "floor_id": {
                    "type": "string"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}